Reorganize (well, add) testing framework.
The new testall.py script in the root of the project scours for test_*.py
files and runs tests therein.

Also change testsupport.py files to run tests in their directories when
run as a script.

Move a couple of REALLY broken tests so that the testall.py script doesn't
find them.