Make appspace/test/test_exec.py pass.
This involved a veritable boatload of fixes:

Hacked opcode_app.py:exec_statement to only use the level of types
yoga we currently support.

Define some methods on issubtype to avoid blowing up quite so
often (Armin probably won't like this :-).

Add has_key() to dicts.

unwrap the code object before execing it in EXEC_STMT (we need to
think how sensible this is as a general policy, but we need to do it
for now).

Deal with the "wrapped defaults" problem for BuiltinModule.compile (we
need to think about this, too).