Add facilities for dynamic loading of builtin modules.
Meta information about the modules provided by the pypy/modules
directory is now encoded in pypy.modules.__init__.

A loader function in BaseObjectSpace reads this information and
dynamically loads the builtin modules. There are facilities to
alter module loading based on object space and other things
such as platform.

The sys and __builtin__ modules are special cased out of the loading
proceedure.

Also included is a simple wrapper to load the platform specific mdules
which do the work of the os module.