Trying to update PyPy's conftest.py for the py refactoring...
With --nomagic it works just fine.  With magic, however, the
reinterpretation hides the original nicely formatted entry; for example,
instead of:

 (app-level) AttributeError: object X has no attribute Y

the line is much longer:

 OperationError:   << [<W_TypeObject(NameError)>: W_StringObject("object X...

The former is generated by AppExceptionInfo.  The latter is produced by
exprinfo.py.  This points to a much more important problem: what if the
reinterpretation raises a different exception?  Arbitrary code may fail for
completely different reasons if it's blindly run a second time...  Displaying
the second exception only in py.test's traceback is a very bad idea :-(