Added the start of a struct module. It is not quite finished yet.
The structure of the code is basically as in structmodule.c. But some python goodies has been used - if that makes the module less RPython I don't know.
TODO:
Find a way to get information on native alignment and sizes.
Convert the CPython tests to py.test
Make the tests pass (all tests pass except a test for a floating point rounding error "SF bug 705836")
Do some more error checking on floats (there is an empty function called "sane_float" which should do the checking.
General cleanup of the code