Armin, look away now :-)
the interpleveldefs for newmodule's ExtModule are now
much more general and flexible. See mixedmodule's
__init__.py for what is possible now.

The original reason for this change was that the sys module
has a lot of actual state ('path', 'modules' ...) which
the previous interpleveldefs' couldn't really specify.
(and overriding getdictvalue() of the sysmodule in particular
seems not nice).

Armin, are you still reading?  Then don't look at
newmodule.getinterpevalloader() :-)