(intermediate check-in, things still break.)
Added proper support for the __builtins__ name in globals, as in CPython.
Now PyFrame.w_builtins has been replaced with PyFrame.builtin, which is the
builtin module to use, instead of a dict (to enable lazy-loading).

Also changed PyCode.co_names_w back to PyCode.co_names, to avoid unwrapping.