issue51 testing
finalizing the module documentation for M0.5 and fixing
a wart for it:  we used to include the contents of
intepreter level sys.path.  Now the implementation
correctly uses PYTHONPATH contents to look up app-level
modules.  The exact import order is now reflected
in the documentation.

Please read/cross-check the new "Modules in PyPy" chapter
in the coding guide.  Putting this issue to testing
to indicate this request.  Note that i introduced
a public query "M0.5-testing" which you can use to
see which issues are currently being tested and how
you can help.