issue48 testing
here is the real commit log regarding the issue48 (describe
all directories).  I played around with ReST and now the
list of directories is a table with an explanation column.
directory references are taken from the common _ref.txt
with an include directive at the end of coding-guide.txt.

The directory-thingie took me at least two hours longer
than i though but i think it's now good enough to be
considered for a prominent location like getting-started.