issue39 resolved
Brain-stormed short mission statement for the front page.  We tried to make it
very self-contained, and very short.  Now a link, "more...", points to an
extended version which gives the main goals.  The links to the existing
projects -- mainly Psyco and Stackless -- are moved in the next paragraph,
as well as to the Related projects section.