Creating a snapshot of the PyPy core, to be used as a stable base to try translation on. translate_pypy.py works on the revision this shapshot was created from.
This branch is NOT MEANT for development.  The translator directories are going to be
removed and replaced by svn:external links to their trunk version, so that working on them
actually makes the trunk progress.  No check-ins really go into the branch, apart from small
fixes that might be needed to make translate_pypy happy (usually as merges from the head).