enabling marshal:
had to initialize buffer in the exception case of mul_string_times.
This was strange, because this was not consequently always needed.
There seems to be some non-determinism.

I found no way to reliably reproduce this. When finally trying
to remove it, the problem came up, again, so I kept it.