issue105 testing
Require explicit care about exception catching in the RTyper.
This is done with an extension of the interface of 'hop' --
which is already quite complicated :-(

* hop.has_implicit_exception(cls) checks if the current high-level
  op is in the scope of a branch catching 'cls', as previously, but
  now also records that the rtyper cares about this case.

* hop.exception_is_here() is called with no argument just before
  a llop is generated.  It means that the llop in question will be
  the one that should be protected by the exception catching.
  If has_implicit_exception() was called before, then
  exception_is_here() verifies that *all* except links in the graph
  have indeed been checked for with an has_implicit_exception().
  This is not verified if has_implicit_exception() has never been
  called -- useful for 'direct_call' and other operations that can
  just raise any exception.

* hop.exception_cannot_occur() is a special case for override:ignore
  meaning that there is, after all, no operation left that could raise
  an exception.  (The RTyper class normally verifies that
  exception_is_here() was really called once for each high-level operation
  that is in the scope of exception-catching links.)

This required adding the correct has_implicit_exception() and
exception_is_here() calls all over the place in the r*.py files. It's possible
that I have overlooked 1-2 untested places.  Indeed, I found a couple of typos
looking around.  It also required more precise 'can_only_throw' attributes in
annotation/binary.py, which is good (but again easy to overlook 1-2 places).