corrected a false comment of mine in floatobject.py.
added some commentary for the tiny _get_odd() function in longobject.py,
which is a bit special because it is the only case where you can use
the lowest bit without taking the sign bit into account.

A general get_bit(a,n) operation would be to expensive to be worth it,
unless we explicitly say it ignores the sign.