support for _alloc_flavor_ in class definitions to allow non-gc allocation strategies for classes.
introduced a flavor (defaults to 'gc') argument to malloc, and flavor_malloc free_non_gc_object
ops

non-gc is assumed if the flavor string doesn't start with 'gc'. In that case a Struct and not a GcStruct
will be used for the instances. There's a common base instance object for such non-gc classes, using
the rclass.NONGCOBJECT Struct.