(pedronis, arigo)
supporting again specialize:memo.  The code is now entierely contained
in one page of specialize.py, which is nice.

However, we have just found out that memoizing with a single PBC argument
is not enough to support multiple spaces, so that nice page will have to
be expanded again...  but probably still only in specialize.py.