(mwh, arigo, pedronis)
RTemoved support for memo functions with two pbcs args, it did
not solve the case we are interested in: Cache.getorbuild.

Introduced instead a new specialisation methodmemo, it does
memoisation wrt both of the 2 supported and expected
arguments and specialises wrt the type of the first argument.

test_call_memoized_cache passes now.