(pedronis, arigo)
As shown by an rtyper test, we need to make MethodDesc a bit more complicated.
Now they have an 'originclassdef' for the class they come from, and (when a
getattr occurs) they have a 'selfclassdef' for the class of the instance that
they were read from.

SomePBC.__init__() now takes care of the corner case that makes the annotator
'assert contains()' unhappy.

Incidentally, this allows us to re-enable a disabled test from test_annrpython.