(pedronis, arigo)
Pushed and pulled obscurity around to make the genc start_new_thread tests
pass.  The idea is to use a HalfConcreteWrapper marker when calling a ll
helper from RPython-level code with a function argument.  Such a function
is meant as a call-back, so it needs particular care, e.g. it should be
turned exactly into a ll function pointer (and not, say, a Void because
it's a single function).