(pedronis, arigo)
Yet Another Refacoring of The Memo Mess (tm).

There are again two phases, discovery of the possible argument values (at
which point the specializer returns the most precise annotation as a result,
but not a complete graph to compute it); and then, just before fixpoint is
reached, the memo tables built so far are "forced" and become graphs that can
be called.  This allows the general N-arguments logic to be written in a
kind-of-sane-but-a-bit-longish way.  It would be easy to add support for Bool
arguments now.