(pedronis, arigo)
Bug, test and fix.  The diff looks short but don't underestimate the
time it took to get it.

In case of generalization, instead of trying to build the next
generalized block "by hand", we now only generalize the redboxes state
and reuse the existing logic to start a new block from this state.