(arre, pedronis)
at least for the cases of "allocation" operations where a null pointer return can only mean
failure and that an exception is being raised, just check for the result of the operation
instead of in the global exception values. This allows in the hint-annotator more control
over avoding virtual container construction (which is compile-time) to be surrounded
by confusing runtime malloc exception checking.