(arre, arigo)
Yellow calls!  The first real progress of the week, but it was worth the
effort.

Turned the dispatch_queue's lists of jitstates into chains (linked lists)
to make it easier to give a type to "list of jitstates": it's the same as
"jitstate".  (It's also more efficient, and less readable.)