(pedronis, arigo)
A more persistent 'debug_assert' operation, with a nice function interface
in objectmodel too.  Unknown overhead :-(

It is compiled to a call to the C macro RPyAssert(), which is by default
#defined to nothing.  To enable, use any of the makefile's "debug"
targets.

Added a lot of debug_assert() in rlist.  Now a debug-compiled pypy-c
should no longer be able to crash obscurely after it popped from an
empty list.