Make box() accept also arguments that inherits from System.Object: this is necessacy for compatibility with PythonNet; consider this case (at RPython level):
  x = ArrayList()
  x.Add(42)
  y = x.get_Item(0)
  y.ToString()

when translated there is no problem, because y's ootype is
NativeInstance(System.Object) e y is a boxed value. However when
running on top of PythonNet the value is automatically unboxed, so y
will contain a plain python int, which of course as no ToString()
method.

To solve, just box() the result of get_Item():

  y = box(x.get_Item(0))