(arre, pedronis, arigo)
Add a flag 'known_nonzero' to PtrRedBox.  This allows the propagation of
the fact that some pointers cannot be NULL, which is essential to get
reasonable code out of exception raising.  This keeps exception-raising
paths separate from regular returns - otherwise the compiler would try
to produce code that uses normally the special "-1" value returned in
case of error!