(pedronis, arigo)
Fixes in the colors in the bytecode dispatch loop.  This now threads
next_instr as a green all around.

Freeze the table used by COMPARE_OP.  Grab the fastlocals_w and make it
virtual.  Add an assert in STORE_FAST to ensure that LOAD_FAST knows in
the common case that it won't fail with UnboundLocalError.