(pedronis, arigo)
Remove the unsafe_call operation because its C equivalent is really unsafe:
passing less arguments than expected in a C call is not really valid
and crashes on some platforms with some optimization levels.
Whacked instead at the stackless transform to avoid the operation.

The principle of conservation of mess means that we had to introduce
another similar operation, though, 'adr_call'.  Don't use it.