Kill the memory simulator and lladdress.
Factor out the parts that compute the GC table into a new module
gctypelayout.py, used both by gctransform/framework and by
gcwrapper.  The latter can be used to test GCs by running them
directly, with the llinterpreter doing the calls to it.

Many tests pass, but it's still work in progress.