(cfbolz, arigo)
Big refactoring.  Every Smalltalk object used as the class of another
Smalltalk object now gets a "mirror" object that only exists in the
interpreter.  The mirror decodes the mind-bogglingly complicated
bitfields and the method dict and stores it in nicely readable RPython
fields.  The internal class field of each Smalltalk object is now
poiting to the mirror directly.  This changes mostly everything done so
far, but should be much more future-proof and cleaner from various
points of view.