(alexander, armin)
Removed the remaining suggested primitives - the stack and stackless manipulation.

We had to resolve circular imports by removing "import objspace" from
pypy/objspace/flow/__init__.py.  This required some other imports to be fixed.

Improved the controllerentry logic for rstack.py.

In translator/c/test/: it's not really possible to test stack_too_big()
directly.  We have other tests in test_typed that checks that we get a
RuntimeError in case of infinite recursion, so that will have to be good
enough.  The issue is that recursive functions automatically contain a
stack_too_big() check, which might be triggered before our explicit one.