Merge of the lazy-write-barrier branch by fijal.
With the generation gc, prebuilt gc objects are now only recorded as
potential static roots when a gc pointer is written into them at
run-time.  This reduces the number of static roots a lot, although
it only gives marginal speed improvements (there seem to be not that
many full collects during a regular pypy-c execution).