Loading nanos has the side-effect of inserting the os module in sys.modules, which causes the whole os.pyc to be embedded as a frozen module inside pypy-c.
The effect was that on Windows, the frozen os.pyc contains a reference to a posix.environ
which is different from the posix.environ which gets initialized on pypy-c startup.

Why only on Windows? Maybe because the posix module is named "nt"; it's possible that the
module is initialized twice, and this should be corrected.

Anyway, it seems better to not embed os.pyc in pypy-c, and let it load the actual file
when it starts.

Now os.environ is correctly initialized on Windows. Phew.