Make PyPy threads compatible with the framework GCs. (svn merge -r55299:55365 svn+ssh://codespeak.net/svn/pypy/branch/gc+thread)
The approach is to create one shadow stack for each thread.
Done by adding three ad-hoc operations that the gctransformer
knows about, gc_thread_*().  This is all a bit fragile, sadly.

pypy/module/thread/os_thread is a bit simplified by replacing
the hack of the raw-malloced Bootstrap instances with a single
prebuilt global Bootstrap instance, and a lock to protect it
as needed.  This should remove all GC-specific issues that where
related to this strange instance.

pypy/module/thread uses rffi even more directly now, relying on
the exact details of the GIL-releasing logic that rffi can put
in external function calls and in callbacks.