Someone, please check:
The "signal" module used to install its hook very early,
when the MixedModule was instantiated:
it was active during all the translation phase,
and the pypysig_poll() function was called many times!

Now we activate it only on module startup, that is, (I hope) when the translated pypy-c starts.

This may be the last remaining usage of ctypes during translation.