First step towards extension modules for Pypy: This test compiles a shared library that uses some space operations, and runs them on top of an *interpreted* object space.
The interesting thing it that it should be possible to use the same
compiled .dll (or .so) on both a translated pypy-c or the interpreted bin/py.py.

I learned a lot about various aspects of the translation. Very hard.
Many aspects of the code could be improved... Comments are warmly welcome!

My next step is to add more space operations, enough to build and
load a module into the pypy interpreter.