Intermediate check-in.  Refactoring the "periodic actions" approach.
This should allow signals and app-level destructors to be handled
promptly, i.e. before the next opcode is run (so far they have to wait
for up to sys.checkinterval opcodes).

The 2nd goal is to decrease the overhead in the fast path of
executioncontext.bytecode_trace() by making the checks here simpler.  It
also seems that a non-threaded pypy-c doesn't even need to increment a
bytecode counter there.