(yoli, antocuni)
(sigh) evil hack to make pypy-{cli,jvm} to translate again on windows; the
problem is that streamio.py assumes that the backend is c, so it freely uses
rffi externals and win32 APIs to implement ftruncate, but of course there is
no chance to make it working on ootype.  It is really a design issue, we need
to be able to provide alternative implementations based on (platform, backend)
at some point.