(pedronis, arigo)
Generate much shorter failure paths.  The idea is to use
generate_quick_failure() to write just a jump to control code which
itself, if ever called, writes the proper failure handling code.  In
this way, for most guards we have only the jump to control code, and for
the few guards that we have actually seen failing at least once we have
the same relatively efficient code as before.