Add tracking of which values are seen by failing GUARD_VALUEs.
This logic is activated by calling make_a_counter_per_value() on a
ResumeGuardDescr.  By default each ResumeGuardDescr uses a single
counter, like before; but if make_a_counter_per_value() is called, then
it will lazily make an instance of ResumeGuardCounters and use that to
track which values are seen.

The algo in ResumeGuardCounters needs improvements, but it should work
as a first draft.