Allow 'whitespace' to form a valid token, when nothing more can match.
This gives a better error position: when given "if ?:", the parser
will return whitespace as a token, and stop on the ?.
Previously, the error would be at the space character,
and one test in interpreter/test/test_syntax.py failed.

This effectively reverts the only effective change in r77650,
which I did not understand correctly at the time.