Main change: allows us to call jitdriver.jit_merge_point() without ever calling jitdriver.can_enter_jit().  In that case the jit_merge_point() plays both roles.  The difference with putting explicitly a can_enter_jit() just before is that such a can_enter_jit() is not seen unless we are closing a loop; in particular, it does not work if we have no loop at all.
Tests and fixes for greenfields.

Implement rlib.jit.jit_debug(), which just causes a JIT_DEBUG
resoperation to show up in the trace of the JIT.

Implement rlib.jit.assert_green(), which fails if the JIT considers
that the argument is not a Const.

When not translated, add the 'FORCE' prefix to the name of resoperations
produced by forcing virtualizables.