(iko, arigo)
A bit unsure, but it seems that thread.locks, implemented on OSX as
pairs of a mutex and a condition variable, don't really reproductibly
survive a fork().  So we do some ugly hack just to reset all of them
after a fork().  Obscure, particularly because CPython doesn't seem
to need it, and we have no clue why.