(iko, rguillebert, arigo)
CPython 2.7 semantics are too messy to follow exactly, e.g.
setuid(-2) works on 32-bit but not on 64-bit.  The exact rules
are rather complicated, and of course the ones we implemented are
a bit different than CPython's.  As a result, we decided to just
accept any 'int', i.e. any C signed long.  This is basically
reverting back to the situation of CPython 2.5, I think.