drop -mmacosx-version-min=10.4 when building on Mac OS X
Just passing -mmacosx-version-min isn't sufficient for ensuring
compatibility with 10.4. In order to make prevent the compiler and
linker from using anything introduced in a later OS, a sysroot should
be specified, pointing to an appropriate SDK.

However, PyPy as-is doesn't even run terribly well on the latest (and most
popular) release of Mac OS X. Supporting older releases amounts to
wishful thinking.

In theory, compiler and linker produce better code when targeting the
latest OS, but any such improvements will probably be minuscule.