Fix the test.  This gives a first approximation of the .NET AddMemoryPressure().  This version is simpler than a counter that needs to be carefully incremented and decremented by the exact same amount.  The idea is to use track_allocation=False to know when a raw malloc is going to be attached to a GC object (good enough for now).  All such raw mallocs make the next major collection occur earlier.  So the major collection is triggered when
    sum( GC object surviving minor collections +
         rawmallocs with track_allocation=False )  >  trigger

The raw mallocs are attached to GC objects with a __del__, which
are never allocated young, so they will survive at least until
the major collection.  But if they survive for longer, they are
ignored for future major collections.  This is again an
approximation, but in the "safe" way.  Indeed, *not* ignoring
them would simply mean having a higher trigger, computed as
(1.82*previous size).