tentative checkin to fix a bug which originates in SVN rev 74676, the merge of the py131 branch.
The problem comes when we call gateway.applevel(..., filename=__file__), which
happens e.g. in pyopcode.py. In this case, we overwrote the content of
linecache.cache[filename], with the result that inspect.getsource() no longer
worked (and e.g. pdb++ was unable to show the source code). The problem got
somehow unnoticed because most of the time __filename__ points to the pyc
file, so the linecache of the source .py file was not overwritten. However,
the problem appears whenever we modify one of these files which contains a
call to applevel(..., filename=__file__).

Fix it by modifying linecache only if the filename is "fake". I'm not sure
whether it's useful at all, but at least it's harmless. The comment says that
it's needed to show tracebacks, but I didn't find any case in which one of
these tracebacks from applevel() code is shown.

Comments welcome :-)