Provide fast paths in find for integer and float strategy lists.
This patch affects "x in l" and "l.index(x)" where l is a list. It leaves the
expected common path (searching for an integer in an integer list; for a
float in a flaot list) unchanged. However, comparisons of other types are
significantly sped up. In some cases, we can use the type of an object to
immediately prove that it can't be in the list (e.g. a user object which
doesn't override __eq__ can't possibly be in an integer or float list) and
return immediately; in others (e.g. when searching for a float in an integer
list), we can convert the input type into a primitive that allows
significantly faster comparisons.

As rough examples, searching for a float in an integer list is approximately
3x faster; for a long in an integer list approximately 10x faster; searching
for a string in an integer list returns immediately, no matter the size of
the list.